Finalizada as tarefas

+++ C#/CS_CODE_/ClassDiary/Program.cs
@@ -7,33 +7,33 @@ namespace Classe
         public static void Main(string[] args){
             System.Console.WriteLine("Bem vindo ao diario de classe!");
 
-            float mediaFinal = 0;
+            decimal mediaFinal = 0;
             System.Console.WriteLine("Qual a nota da primeira prova? (Por favor, use o ponto como separador decimal, exemplo: 9.7)");
-            float nota1 = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
+            decimal nota1 = decimal.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
 
             mediaFinal = nota1;
             System.Console.WriteLine($"Qual a nota da segunda prova? (Por favor, use o ponto como separador decimal, exemplo: 9.7)\nMédia atual: {mediaFinal.ToString("f2")} pontos");
-            float nota2 = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
+            decimal nota2 = decimal.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
 
             mediaFinal = (nota1+nota2)/2.0F;
             System.Console.WriteLine($"Qual a nota da terceira prova? (Por favor, use o ponto como separador decimal, exemplo: 9.7)\nMédia atual: {mediaFinal.ToString("f2")} pontos");
-            float nota3 = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
+            decimal nota3 = decimal.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
 
             mediaFinal = (nota1+nota2+nota3)/3.0F;
             System.Console.WriteLine($"Qual a nota da quarta prova? (Por favor, use o ponto como separador decimal, exemplo: 9.7)\nMédia atual: {mediaFinal.ToString("f2")} pontos");
-            float nota4 = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
+            decimal nota4 = decimal.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
 
             mediaFinal =(nota1+nota2+nota3+nota4)/4.0F ;
 
             
-            float notaMinimaRequerida = 7.0F;
-            float notaRecuperacao = default(float);
+            decimal notaMinimaRequerida = 7.0F;
+            decimal notaRecuperacao = default(decimal);
             if (mediaFinal>=notaMinimaRequerida){
                 System.Console.WriteLine($"É com prazer que informamos que o aluno foi aprovado!\nMédia final de {mediaFinal.ToString("f2")} pontos.");
                 System.Environment.Exit(0);
             }else{
                 System.Console.WriteLine($"\nO aluno tem direito a uma prova de recuperação! Média final atual: {mediaFinal.ToString("f2")} pontos.\nQual a nota da prova de recuperação? (Por favor, use o ponto como separador decimal, exemplo: 9.7)");
-                notaRecuperacao = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
+                notaRecuperacao = decimal.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
                 mediaFinal = (mediaFinal+notaRecuperacao)/2.0F;
             }
